requirements.txt
Django
PyMySQL
djangorestframework
djangorestframework-filters
markdown
django-filter
scrapy
urllib3
requests
tensorflow
Pillow
pygments
Celery
celery[redis]
coreapi